small refactor for db connection, start work on score db
